Add LuaCATS annotations to Chunks
This was disappointingly hard to do, as I basically needed to bend a lot
to a very specific way to annotate if I wanted to have no errors or
warnings, and code completion sort of worked, but at least the code is
unchanged.

As a side effect, removed the `shared_new` function (which wasn't that
much shared, and it was quite trivial anyway), and clarified a bit the
unit tests.

With the new docs and the better tests, I hope I can properly remember
what's the intention of code better. Which it was definitely lost from
short term memory after not touching the project for long. Now it needs
to be done for everything else.