Improve the unit tests
Replace `setup` with `before_each`. Is not strictly needed for all
cases, but for consistency.

Get rid of the stub. It's cool, but while trying to improve the test to
clear the stub after the test has done its thing, I've realized it's not
super documented, and I don't even know why I went with using it when I
totally don't need to `print` at all on the tests to log how many
iterations have been done.