Some small tweaks and add an auto test
Renamed the `text` function to `to_code` for now (I'll figure out a
better name later, if needed, as it's not necessarily going to be
public). But at least this name is easy to search and replace later.

Added a few more comments and docs, as returning back to the project
is harder than it should be.

Also, added an automated integration test to the `to_code` function.