Implement the `execute` function
This wraps the given function in an environment, which is also merged
with the global one, and providing the `escape` and `tostring`
functions.

This is a bit different from etlua, at least for now. I'm not sure about
the reasons for which etlua passes as parameters the buffer and the two
functions to convert to string and to escape. Perhaps passing them as
locals improves a bit performance as they don't need to be looked up in
the environment. It seems etlua also might be avoiding `table.insert` in
order to avoid any lookup into the global environment. I've not done so,
as I think it's way clearer to write an insert as `table.insert`.

But since all the values of the environment need to be looked up at some
point in order to provide values, I'm not sure it's worth the effort. I
am considering going the exact opposite way, and inject the buffer as
well, as long as it's not an issue with concatenating buffers. This
would make everything more coherent: everything is injected.

Otherwise, I might change course and instead go to inject nothing, and
pass the functions and buffer as parameter, like etlua does.