The big newline-handling refactor
When I started this project, I wanted to simplify as much as possible
the feature set, while still retaining being expressive and clear in the
templates. And I noticed that in my etlua templates I used the newline
trimming all the time. That's why I initially thought that this trimming
was better as the default, and no specific markup for asking for it
explicitly was necessary because adding a new line in text mode would be
better.

But I already had the support for the "this whole line is code"
delimiter as something that was necessary to avoid being so verbose when
making a simple loop. And that markup already avoided adding the end of
line to the output.

So I noticed that the reverse is actually more sensible. While I still
don't rule out explicit markup to the closing delimiter to specify an
end of line trimming (and perhaps changing the default), right now I'll
go with the line-of-code delimiter being the one that always skips
newlines, and the closing delimiter NOT trimming the newline after it.
It's actually possible to trim the newline while keeping the markup
almost the same, by moving the closing delimiter to its own line. Like:

    This is one line in the output, <%= name
    %>!

While still allowing this being one line per bullet list, with the
simplest possible markup and with the newlines in the right place:

    My list of items:
    % for _, item in ipairs(list) do
    * <%= item %>
    % end

So I've reworked a lot the parse function, and I've added extensive unit
tests with input/output pairs in a "samples" directory (and I've added
an Asciidoc file there that showcases all the tests, which could be also
helpful as documentation). More granular tests at the Parser unit test
level are also added (and the old tests are adjusted).

The Parser:parse function now:
* Fixes the issue of an output value (Chunks.RAW or Chunks.ESCAPE) being
  spread into multiple lines. The parser now groups it into one chunk,
  so there is only one `tostring` call.
* Simplifies the logic a lot, grouping the different cases as much as
  possible, and flattening the conditions.
* Changes the newline logic to the new design.